requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
chardet==3.0.4
docutils==0.19
emoji==2.1.0
idna==2.10
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
lxml==4.9.1
Pygments==2.13.0
pypiwin32==223
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pywin32==304
requests==2.24.0
six==1.16.0
snowballstemmer==2.2.0
tornado==6.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.25.11


